правка багов и новый функционал
- исправлены ошибки с падением сервиса
- добавлены обработки ошибок 404
- замена выдачи 401 в проверке прав у юзера на 204
- возврат ошибки 403 при отсутствии прав
- добавление возможности работы с командной строкой в app.py
- проверка прав по логину и по имени роли
- добавление супер-пользователя при старте приложения
- добавление функций получения id пользователя по email и email по id
- исправление имени функции обновления refresh токена, чтобы избавиться от непреднамеренной рекурсии
- исправление проверки id при смене данных пользователя - менять данные другим пользователям может только админ

+++ src/api/v1/role.py
@@ -3,14 +3,15 @@ from api.v1.error_messages import APIErrors
 from flask import Blueprint, request, Response, jsonify
 from flask_jwt_extended import jwt_required, get_jwt_identity
 
-role_routes = Blueprint("role", __name__, url_prefix="/role")
+role_routes = Blueprint("role_routes", __name__, url_prefix="/role")
 
 
 @role_routes.route("/", methods=["GET", "POST"])
 @jwt_required()
 def role():
     identy = get_jwt_identity()
-    service_role.check_user_role_by_email(identy, "admin")
+    if service_role.check_user_role_by_email(identy, "admin") != "OK":
+        return Response(status=403, mimetype="application/json")
     if request.method == "POST":
         request_data = request.get_json()
         role_name = request_data["name"]
@@ -33,22 +34,26 @@ def role():
 @jwt_required()
 def role_crud(role_id):
     identy = get_jwt_identity()
-    service_role.check_user_role_by_email(identy, "admin")
+    if service_role.check_user_role_by_email(identy, "admin") != "OK":
+        return Response(status=403, mimetype="application/json")
     if request.method == "PUT":
         request_data = request.get_json()
         role_name = request_data["name"]
         change = service_role.change_role(role_id, role_name)
         if change == APIErrors.ROLE_OR_USER_NOT_FOUND:
             return Response(status=404, mimetype="application/json")
+        if change == APIErrors.ROLE_NAME_ALREADY_TAKEN:
+            return Response(status=409, mimetype="application/json")
         return Response(status=200, mimetype="application/json")
 
     if request.method == "DELETE":
-        service_role.delete_role(role_id)
+        res = service_role.delete_role(role_id)
+        if res == APIErrors.ROLE_NOT_FOUND:
+            return Response(status=404, mimetype="application/json")
         return Response(status=200, mimetype="application/json")
 
     if request.method == "GET":
         role = service_role.role_by_id(role_id)
-        print(role)
         if role == APIErrors.ROLE_OR_USER_NOT_FOUND:
             return Response(status=400, mimetype="application/json")
         return role
@@ -59,7 +64,8 @@ def role_crud(role_id):
 @jwt_required()
 def user_role_crud(user_id, role_id):
     identy = get_jwt_identity()
-    service_role.check_user_role_by_email(identy, "admin")
+    if service_role.check_user_role_by_email(identy, "admin") != "OK":
+        return Response(status=403, mimetype="application/json")
     if request.method == "PUT":
         assign = service_role.assign_user_role(
             user_id=user_id,
@@ -85,6 +91,10 @@ def user_role_crud(user_id, role_id):
                    methods=["GET"])
 @jwt_required()
 def user_role_get(user_id, role_id):
+    identy = get_jwt_identity()
+    if service_role.check_user_role_by_email(identy, "role_checker") != "OK"\
+            and service_role.check_user_role_by_email(identy, "admin") != "OK":
+        return Response(status=403, mimetype="application/json")
     check = service_role.check_user_role(
         user_id=user_id,
         role_id=role_id
@@ -92,6 +102,6 @@ def user_role_get(user_id, role_id):
     if check == APIErrors.ROLE_OR_USER_NOT_FOUND:
         return Response(status=404, mimetype="application/json")
     if check == APIErrors.USER_DOESNT_HAVE_ROLE:
-        return Response(status=400, mimetype="application/json")
+        return Response(status=204, mimetype="application/json")
     if check == "OK":
         return Response(status=200, mimetype="application/json")
